Improved TimeSlider
